layout: add commit hash to the footer of each page
Add the abbreviated hash of the latest commit that modified the current
page to the footer, as a way to subtly insert what the last modification
performed was.

This looks nicer (and edgier) than putting in a timestamp for the latest
edit, but would require people reading to have some knowledge of Git
before they can get the information they want.

GitInfo was enabled in the site's config.